Refactor detector version sourcing into shared module
