+++ src/index.test.ts
@@ -1,16 +1,11 @@
-import fs from 'fs'
-import path from 'path'
+import { getDefaultSnapshotDetector } from './detector'
 
-import { getDefaultSnapshotDetector, getProjectVersion } from './index'
+const packageJson = require('../package.json') as {
+  version: string
+}
 
 describe('default snapshot detector', () => {
   test('version matches project version', () => {
-    const packageJsonPath = path.resolve(__dirname, '..', 'package.json')
-    const packageJson = JSON.parse(fs.readFileSync(packageJsonPath, 'utf8')) as {
-      version: string
-    }
-
-    expect(getProjectVersion()).toEqual(packageJson.version)
     expect(getDefaultSnapshotDetector().version).toEqual(packageJson.version)
   })
 })
